+++ CHANGELOG.md
@@ -14,6 +14,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Fixed
 
 - Small bug in pressure drop lateral pipe.
+- Small bug in pressure drop calculation of separatus.
 
 ## [2.3.3] - 2025-05-28